Fix format warnings from debug output
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17594 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/ip-mib/ipAddressTable/ipAddressTable_data_access.c
@@ -395,7 +395,7 @@ ipAddressTable_container_load(netsnmp_container *container)
     }
 
     DEBUGMSGT(("verbose:ipAddressTable:ipAddressTable_cache_load",
-               "%d records\n", CONTAINER_SIZE(container)));
+               "%lu records\n", (unsigned long)CONTAINER_SIZE(container)));
 
     return MFD_SUCCESS;
 }
